is_s3_generic {roxygen2}	R Documentation

Determine if a function is an S3 generic or S3 method.

Description

is_s3_generic compares name to .knownS3Generics and .S3PrimitiveGenerics, then looks at the function body to see if it calls UseMethod().

is_s3_method builds names of all possible generics for that function and then checks if any of them actually is a generic.

Usage

is_s3_generic(name, env = parent.frame())

is_s3_method(name, env = parent.frame())

Arguments

name
Name of function.

env
Base environment in which to look for function definition.

[Package roxygen2 version 7.1.2 Index]