object_format {roxygen2}	R Documentation

Default format for data

Description

This function is called to generate the default "Format" section for each data object. The default implementation will return the class and dimension information.

Usage

object_format(x)

Arguments

x
A data object

Value

A character value with valid Rd syntax, or NULL.

[Package roxygen2 version 7.1.2 Index]