rd_roclet {roxygen2}	R Documentation

Roclet: make Rd files.

Description

This roclet is the workhorse of roxygen, producing the Rd files that document that functions in your package.

Usage

rd_roclet()

See Also

vignette("rd", package = "roxygen2")

Other roclets: S3method, export, exportClass, exportMethod, import, importClassesFrom, importFrom, importMethodsFrom, namespace_roclet

[Package roxygen2 version 4.1.0 Index]