tags-index-crossref {roxygen2}	R Documentation

Tags for indexing and cross-references

Description

Learn the full details in vignette('index-crossref').

Key tags:

@aliases ${1:alias}: Add additional aliases to the topic. Use NULL to suppress the default alias automatically generated by roxygen2.

@concept ${1:concept}: Add additional keywords or phrases to be included in the help.search() index. Each @concept should be a single term or phrase.

@family ${1:family name}: Generate @seealso entries to all other functions in family name.

@keywords ${1:keyword}: Add a standardised keyword, indexed by help.search(). These are generally not useful apart from @keywords internal which flags the topic as internal and removes from topic indexes.

@references ${1:reference}: Pointers to the related literature. Usually formatted like a bibliography.

@seealso [${1:func}()]: Link to other related functions or urls. Usually a sentence or two, or a bulleted list.

Other less frequently used tags:

@backref ${1:path}: Manually override the backreference that points from the .Rd file back to the source .R file. Only needed when generating code.

Usage

#' @aliases ${1:alias}
#' @backref ${1:path}
#' @concept ${1:concept}
#' @family ${1:family name}
#' @keywords ${1:keyword}
#' @references ${1:reference}
#' @seealso [${1:func}()]

See Also

Other documentation tags: tags-rd, tags-rd-other, tags-reuse

[Package roxygen2 version 7.3.3 Index]