tags-rd-other {roxygen2}	R Documentation

Tags for documenting datasets and classes

Description

Learn the full details in vignette('rd-other').

Key tags:

@field ${1:name} ${2:description}: Describe a R6 or refClass field.

@format ${1:description}: Describe the type/shape of a dataset. If the dataset is a data frame, include a description of each column. If not supplied, will be automatically generated by object_format().

@method ${1:generic} ${2:class}: Force a function to be recognised as an S3 method. This affects the default usage and the NAMESPACE directive produced by @export. Only needed if automatic detection fails.

@slot ${1:name} ${2:description}: Describe the slot of an S4 class.

@source ${1:description}: Describe where the dataset came from. Provide a link to the original source (if possible) and briefly describe any manipulation that you performed when importing the data.

Usage

#' @field ${1:name} ${2:description}
#' @format ${1:description}
#' @method ${1:generic} ${2:class}
#' @slot ${1:name} ${2:description}
#' @source ${1:description}

See Also

Other documentation tags: tags-index-crossref, tags-rd, tags-reuse

[Package roxygen2 version 7.3.3 Index]