load_options {roxygen2}	R Documentation

Load roxygen2 options

Description

Options can be stored in the Roxygen field of the DESCRIPTION, or in man/roxygen/meta.R. In either case, the code is parsed and evaluated in a child of the base environment. Call roxy_meta_get() to access current option values from within tag and roclet methods.

Usage

load_options(base_path = ".")

roxy_meta_get(key = NULL, default = NULL)

Arguments

base_path
Path to package.

Details

Options in man/roxygen/meta.R override those present in DESCRIPTION.

[Package roxygen2 version 7.0.0 Index]