object {roxygen2}	R Documentation

Constructors for S3 object to represent R objects.

Description

These objects are usually created by the parsers, but it is also useful to generate them by hand for testing.

Usage

object(value, alias, type)

Arguments

value
The object itself.

alias
Alias for object being documented, in case you create a generator function with different name.

[Package roxygen2 version 7.1.2 Index]