rd_roclet {roxygen2}	R Documentation

Roclet: make Rd files.

Description

This roclet is the workhorse of roxygen, producing the .Rd files that R uses to document that functions, datasets, packages, classes, and other objects.

See vignette("rd") for details.

Generally you will not call this function directly but will instead use roxygenise() specifying the rd roclet

Usage

rd_roclet()

See Also

Other roclets: namespace_roclet(), vignette_roclet()

Examples

#' The length of a string (in characters)
#'
#' @param x String input character vector
#' @return An integer vector the same length as `x`.
#'   `NA` strings have `NA` length.
#' @seealso [nchar()]
#' @export
#' @examples
#' str_length(letters)
#' str_length(c("i", "like", "programming", NA))
str_length <- function(x) {
}

[Package roxygen2 version 7.1.2 Index]