Roclet: make Rd files

This roclet is the workhorse of roxygen2, producing the .Rd files that R uses to document functions, datasets, packages, classes, and more. See vignette("rd") for details.

Generally you will not call this function directly but will instead use roxygenise() specifying the rd roclet.

Usage

rd_roclet()

See also

Examples

#' The length of a string (in characters)
#'
#' @param x A character vector.
#' @returns An integer vector the same length as `x`.
#'   `NA` strings have `NA` length.
#' @seealso [nchar()]
#' @export
#' @examples
#' str_length(letters)
#' str_length(c("i", "like", "programming", NA))
str_length <- function(x) {
}