roxy_tag {roxygen2}	R Documentation

Parsing tags.

Description

roxy_tag constructs a tag object, and roxy_tag_warning makes an informative warning using the location information stored in the tag. The remainder of the tag functions parse the tag value, convert a string into a richer R object, or providing informative warnings and returning valid if the value is invalid.

Usage

roxy_tag(tag, val, file = "", line = 0)

roxy_tag_warning(x, ...)

tag_value(x)

tag_inherit(x)

tag_name(x)

tag_two_part(first, second, required = TRUE)

tag_name_description(x)

tag_words(min = 0, max = Inf)

tag_words_line(x)

tag_toggle(x)

tag_code(x)

tag_examples(x)

tag_markdown(x)

tag_markdown_restricted(x)

Arguments

tag
Tag name

val
Tag value. When read from the file, this will be a string, but after parsing can be a more complicated structure (typically a character vector, but sometimes a list).

file, line
Location of the tag

first, second
Name of first and second parts of two part tags

required
Is the second part required (TRUE) or can it be blank (FALSE)?

min, max
Minimum and maximum number of words

Details

Two exceptions to the rule are tag_words and tag_two_part, which are tag parsing generator functions.

[Package roxygen2 version 6.1.1 Index]