roxy_tag {roxygen2}	R Documentation

roxy_tag S3 constructor

Description

roxy_tag() is the constructor for tag objects. roxy_tag_warning() generates a warning that gives the location of the tag.

Usage

roxy_tag(tag, raw, val = NULL, file = NA_character_, line = NA_integer_)

roxy_tag_parse(x)

roxy_tag_warning(x, ...)

Arguments

tag
Tag name. Arguments starting with . are reserved for internal usage.

raw
Raw tag value, a string.

val
Parsed tag value, typically a character vector, but sometimes a list. Usually filled in by tag_parsers

file, line
Location of the tag

x
A tag

Methods

Define a method for roxy_tag_parse to support new tags. See tag_parsers for more details.

[Package roxygen2 version 7.1.2 Index]