roxygen2-package {roxygen2}	R Documentation

roxygen2: In-Line Documentation for R

Description

Generate your Rd documentation, 'NAMESPACE' file, and collation field using specially formatted comments. Writing documentation in-line with code makes it easier to keep your documentation up-to-date as your requirements change. 'Roxygen2' is inspired by the 'Doxygen' system for C++.

Details

See vignette("roxygen2", package = "roxygen2") for an overview of the package, vignette("rd", package = "roxygen2") for generating documentation, and vignette("namespace", package = "roxygen2") for generating the namespace specification.

If you have existing Rd files, check out the Rd2roxygen package for a convenient way of converting Rd files to roxygen comments.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com [copyright holder]

Authors:

Peter Danenberg pcd@roxygen.org [copyright holder]

Manuel Eugster [copyright holder]

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://github.com/klutometis/roxygen

Report bugs at https://github.com/klutometis/roxygen/issues

[Package roxygen2 version 6.1.1 Index]