roxygen2-package {roxygen2}	R Documentation

In-Source Documentation for R

Description

A 'Doxygen'-like in-source documentation system for Rd, collation, and 'NAMESPACE' files.

Details

If you have existing Rd files, check out the Rd2roxygen package for a convenient way of converting Rd files to roxygen comments.

Author(s)

Hadley Wickham h.wickham@gmail.com, Peter Danenberg pcd@roxygen.org, Manuel Eugster Manuel.Eugster@stat.uni-muenchen.de

Maintainer: Hadley Wickham h.wickham@gmail.com

See Also

See vignette("roxygen2", package = "roxygen2") for an overview of the package, vignette("rd", package = "roxygen2") for generating documentation, and vignette("namespace", package = "roxygen2") for generating the namespace specification.

Examples

## Not run: roxygenize('pkg')

[Package roxygen2 version 5.0.1 Index]