source_package {roxygen2}	R Documentation

Source all files in a package.

Description

This is a simple attempt to load code in a package used by roxygenize. It will work with simple packages, but fail if there are compiled files, data files, etc. In that case, it's better to use document().

Usage

source_package(path)

Arguments

path
Path to a package.

Value

An environment, into which all R files in the directory were sourced.

[Package roxygen2 version 5.0.1 Index]