vignette_roclet {roxygen2}	R Documentation

Re-build outdated vignettes.

Description

This rebuilds outdated vignettes with tools::buildVignette. By default, it will rebuild all vignettes if the source file is newer than the output pdf or html. (This means it will automatically re-build the vignette if you change the vignette source, but not when you change the R code). If you want finer control, add a Makefile to vignettes/ and roxygen2 will use that instead.

Usage

vignette_roclet()

Details

To prevent RStudio from re-building the vignettes again when checking your package, add --no-build-vignettes to the "Build Source Package" field in your project options.

See Also

Other roclets: namespace_roclet(), rd_roclet()

[Package roxygen2 version 7.1.2 Index]